requirements.txt
anthropic
pydantic
pandas
streamlit
pytest
python-dotenv